Create Sophyia_number_cardinality.py
This program is a simple coding assessment I came across in an interview; it returns odd if the value inputted is not divisible by two and even if it is divisible by 2. For values ending with zero, it returns zero as a string, and the same goes for numbers ending with five. In addition, I included an exception handler to catch errors from user input.

I hope it helps someone.